+++ src/Renderer/Sprite.h
@@ -22,7 +22,7 @@ namespace RenderEngine
 		{
 			CFrameDescription(const glm::vec2& _leftBottomUV,
 				const glm::vec2& _rightTopUV,
-				const uint64_t _duration) :
+				const double _duration) :
 				leftBottomUV(_leftBottomUV),
 				rightTopUV(_rightTopUV),
 				duration(_duration)
@@ -30,7 +30,7 @@ namespace RenderEngine
 
 			glm::vec2 leftBottomUV;
 			glm::vec2 rightTopUV;
-			uint64_t duration;
+			double duration;
 		};
 
 
@@ -47,7 +47,7 @@ namespace RenderEngine
 			const float fLayer = 0.f, const size_t frameId = 0) const;
 
 		void insertFrames(std::vector<CFrameDescription> frameDescriptions);
-		uint64_t getFrameDuration(const size_t frameId) const;
+		double getFrameDuration(const size_t frameId) const;
 		size_t getFramesCount() const;
 
 	protected:
